requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
discord.py==1.5.1
async-timeout==3.0.1
bs4==0.0.1
requests==2.24.0
matplotlib==3.3.3
seaborn==0.11.1
psycopg2==2.8.6
ujson==4.0.2